small fixes

+++ main.c
@@ -81,10 +81,10 @@ double testPosixIO(const char* outFile, const size_t ost, const size_t size)
     fprintf(stderr, "Can't allocate buffer of size %zu\n", fileSize);
     exit(-1);
   }
-  // TODO: fill up the file content
+  // fill up the file content
   int x1 = rand();
   int x2 = rand();
-  for (int *p = fileContent; p < fileContent+fileSize; p+= 2) {
+  for (int *p = (int)fileContent; p < (int)(fileContent+fileSize); p+= 2) {
     *p = x1;
     *(p+1) = x2;
     x1++;
